requirements.txt
praw
pandas
sqlalchemy
python-dotenv
psycopg2-binary